Haaaaaaaaands

+++ src/Physics.h
@@ -62,8 +62,8 @@ Vector operator+ (Vector, Vector);
 std::ostream& operator<<(std::ostream&, Vector);
 
 struct Rect {
-	float x, y, w, h;
-	Rect(float x, float y, float w, float h) : x(x), y(y), w(w), h(h) {
+	int x, y, w, h;
+	Rect(int x, int y, int w, int h) : x(x), y(y), w(w), h(h) {
 
 	}
 	Rect() : x(0), y(0), w(0), h(0) {
